PIMA Indian Diabetes prediction
The Pima Indians diabeties data has 768 instances, each instance having 8 attributes with which the patients were medically tested upon. Using the data of the tested patients, based on the 8 attributes, we use different classfication ML techniques to predict if the patient is actually a diabetic or not.
The attributes on which the patients were tested are:
1)Pregnencies
2)Plasma glucose concentration a 2 hours in an oral glucose tolerance test
3)Diastolic blood pressure (mm Hg)
4)Triceps skin fold thickness (mm)
5)2-Hour serum insulin (mu U/ml)
6)BMI: Body mass index (weight in kg/(height in m)^2)
7)Diabetes pedigree function
8)Age (In years)